Decide if users are in the right place based on the clipboard data
The function regularly reads the clipboard data and then decides if the user is
in the correct place for the current station. If so, the celebration can be
triggered. If not, an alternative scenario can be set up.

Currently classification_data.csv is a dummy file. It needs to be filled
with the real data in the format Id,bulge,bar.